refactor: reorganize adaptive testing tests into separate files
- Deleted the monolithic test file for adaptive testing and distributed its contents across multiple new files for better organization and maintainability.
- Introduced new test files: test_evaluation.py, test_responses.py, test_tests.py, test_topics.py, and conftest.py to encapsulate related test functionalities.
- This restructuring enhances clarity and allows for easier navigation and modification of tests related to adaptive testing functionalities.
