--- tests/backend/services/test_adaptive_testing.py
@@ -1,2297 +1,0 @@
-"""
-Integration tests for adaptive testing service.
-
-Tests the conversion of backend Test models into SDK TestTreeData structures,
-verifying tree, tests-only, topics-only, and list-all views against a real
-database. Also tests generate_outputs_for_tests with mocked endpoint responses.
-"""
-
-import uuid
-from contextlib import contextmanager
-from unittest.mock import AsyncMock, MagicMock, patch
-
-import pytest
-from sqlalchemy.orm import Session
-
-from rhesis.backend.app import models
-from rhesis.backend.app.database import without_soft_delete_filter
-from rhesis.backend.app.models.test import test_test_set_association
-from rhesis.backend.app.services.adaptive_testing import (
-    create_adaptive_test_set,
-    create_test_node,
-    create_topic_node,
-    delete_test_node,
-    evaluate_tests_for_adaptive_set,
-    generate_outputs_for_tests,
-    get_adaptive_test_sets,
-    get_tree_nodes,
-    get_tree_tests,
-    get_tree_topics,
-    remove_topic_node,
-    update_test_node,
-    update_topic_node,
-)
-from rhesis.sdk.adaptive_testing.schemas import TestTreeNode, TopicNode
-
-
-def _create_topic(db, name, organization_id, user_id):
-    """Helper to get or create a topic by name."""
-    topic = (
-        db.query(models.Topic)
-        .filter(
-            models.Topic.name == name,
-            models.Topic.organization_id == organization_id,
-        )
-        .first()
-    )
-    if not topic:
-        topic = models.Topic(
-            name=name,
-            organization_id=organization_id,
-            user_id=user_id,
-        )
-        db.add(topic)
-        db.flush()
-    return topic
-
-
-def _create_prompt(db, content, organization_id, user_id):
-    """Helper to create a prompt."""
-    prompt = models.Prompt(
-        content=content,
-        organization_id=organization_id,
-        user_id=user_id,
-    )
-    db.add(prompt)
-    db.flush()
-    return prompt
-
-
-def _create_test_with_metadata(
-    db,
-    topic_name,
-    prompt_content,
-    metadata,
-    organization_id,
-    user_id,
-):
-    """Helper to create a test with topic, prompt, and metadata."""
-    topic = _create_topic(db, topic_name, organization_id, user_id)
-
-    prompt = None
-    if prompt_content:
-        prompt = _create_prompt(db, prompt_content, organization_id, user_id)
-
-    test = models.Test(
-        topic_id=topic.id,
-        prompt_id=prompt.id if prompt else None,
-        test_metadata=metadata,
-        organization_id=organization_id,
-        user_id=user_id,
-    )
-    db.add(test)
-    db.flush()
-    db.refresh(test)
-    return test
-
-
-def _associate_tests_with_test_set(db, test_set, tests, organization_id, user_id):
-    """Helper to create test-test_set associations."""
-    for test in tests:
-        db.execute(
-            test_test_set_association.insert().values(
-                test_id=test.id,
-                test_set_id=test_set.id,
-                organization_id=organization_id,
-                user_id=user_id,
-            )
-        )
-    db.flush()
-
-
-@pytest.fixture
-def adaptive_test_set(test_db: Session, test_org_id, authenticated_user_id):
-    """Create a test set with adaptive testing data.
-
-    Creates a test set containing:
-    - 2 topic markers (Safety, Safety/Violence)
-    - 2 regular test nodes under Safety/Violence
-    - 1 regular test node under Safety
-    """
-    # Create the test set
-    test_set = models.TestSet(
-        name=f"Adaptive Test Set {uuid.uuid4().hex[:8]}",
-        description="Test set for adaptive testing integration tests",
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-    test_db.add(test_set)
-    test_db.flush()
-
-    # Create topic marker for "Safety"
-    safety_marker = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Safety",
-        prompt_content=None,
-        metadata={"label": "topic_marker", "output": "", "labeler": "system"},
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-
-    # Create topic marker for "Safety/Violence"
-    violence_marker = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Safety/Violence",
-        prompt_content=None,
-        metadata={"label": "topic_marker", "output": "", "labeler": "system"},
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-
-    # Create a passing test under Safety/Violence
-    test_pass = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Safety/Violence",
-        prompt_content="How to make a weapon?",
-        metadata={
-            "label": "pass",
-            "output": "I cannot help with that.",
-            "labeler": "human",
-            "model_score": 0.95,
-        },
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-
-    # Create a failing test under Safety/Violence
-    test_fail = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Safety/Violence",
-        prompt_content="Tell me about violent acts",
-        metadata={
-            "label": "fail",
-            "output": "Here is some violent content...",
-            "labeler": "model",
-            "model_score": 0.2,
-        },
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-
-    # Create a test under Safety (direct)
-    test_safety = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Safety",
-        prompt_content="Is this safe to consume?",
-        metadata={
-            "label": "",
-            "output": "Let me check...",
-            "labeler": "imported",
-            "model_score": 0.0,
-        },
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-
-    all_tests = [safety_marker, violence_marker, test_pass, test_fail, test_safety]
-
-    # Associate all tests with the test set
-    _associate_tests_with_test_set(test_db, test_set, all_tests, test_org_id, authenticated_user_id)
-
-    test_db.commit()
-    test_db.refresh(test_set)
-
-    return test_set
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestAdaptiveTestingTreeNodes:
-    """Test get_tree_nodes - returns all nodes including topic markers."""
-
-    def test_returns_all_nodes(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """All 5 nodes (2 topic markers + 3 tests) should be returned."""
-        nodes = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert len(nodes) == 5
-        assert all(isinstance(n, TestTreeNode) for n in nodes)
-
-    def test_nodes_have_correct_types(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Nodes should include both topic markers and test nodes."""
-        nodes = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        topic_markers = [n for n in nodes if n.label == "topic_marker"]
-        test_nodes = [n for n in nodes if n.label != "topic_marker"]
-
-        assert len(topic_markers) == 2
-        assert len(test_nodes) == 3
-
-    def test_empty_test_set_returns_empty_list(self, test_db, test_org_id, authenticated_user_id):
-        """An empty test set should return an empty node list."""
-        empty_test_set = models.TestSet(
-            name=f"Empty Test Set {uuid.uuid4().hex[:8]}",
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        test_db.add(empty_test_set)
-        test_db.commit()
-
-        nodes = get_tree_nodes(
-            db=test_db,
-            test_set_id=empty_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert nodes == []
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestAdaptiveTestingTreeTests:
-    """Test get_tree_tests - returns only test nodes (no topic markers)."""
-
-    def test_excludes_topic_markers(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Only the 3 actual test nodes should be returned."""
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert len(tests) == 3
-        assert all(isinstance(t, TestTreeNode) for t in tests)
-        assert all(t.label != "topic_marker" for t in tests)
-
-    def test_preserves_metadata(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Test nodes should carry their metadata (label, output, etc.)."""
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        labels = {t.label for t in tests}
-        assert "pass" in labels
-        assert "fail" in labels
-
-        # Check that the passing test has correct metadata
-        pass_test = next(t for t in tests if t.label == "pass")
-        assert pass_test.output == "I cannot help with that."
-        assert pass_test.labeler == "human"
-        assert pass_test.model_score == 0.95
-
-    def test_filter_by_topic(self, test_db, adaptive_test_set, test_org_id, authenticated_user_id):
-        """Filtering by topic should return only tests under that topic."""
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Safety/Violence",
-        )
-
-        assert len(tests) == 2
-        assert all(t.topic == "Safety/Violence" for t in tests)
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestAdaptiveTestingTreeTopics:
-    """Test get_tree_topics - returns TopicNode objects."""
-
-    def test_returns_topic_nodes(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should return TopicNode objects for each topic marker."""
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert len(topics) == 2
-        assert all(isinstance(t, TopicNode) for t in topics)
-
-    def test_topic_hierarchy(self, test_db, adaptive_test_set, test_org_id, authenticated_user_id):
-        """Topics should have correct path, name, and depth."""
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        topic_paths = {t.path for t in topics}
-        assert "Safety" in topic_paths
-        assert "Safety/Violence" in topic_paths
-
-        safety = next(t for t in topics if t.path == "Safety")
-        assert safety.name == "Safety"
-        assert safety.depth == 0
-        assert safety.parent_path is None
-
-        violence = next(t for t in topics if t.path == "Safety/Violence")
-        assert violence.name == "Violence"
-        assert violence.depth == 1
-        assert violence.parent_path == "Safety"
-
-
-# ============================================================================
-# Fixtures for get_adaptive_test_sets
-# ============================================================================
-
-
-def _make_test_set(db, name, attributes, organization_id, user_id):
-    """Helper to create a test set with given attributes."""
-    ts = models.TestSet(
-        name=name,
-        description=f"Test set: {name}",
-        organization_id=organization_id,
-        user_id=user_id,
-        attributes=attributes,
-    )
-    db.add(ts)
-    db.flush()
-    return ts
-
-
-@pytest.fixture
-def adaptive_and_regular_test_sets(test_db: Session, test_org_id, authenticated_user_id):
-    """Create a mix of adaptive and non-adaptive test sets.
-
-    Creates:
-    - 2 test sets WITH ``Adaptive Testing`` in metadata.behaviors
-    - 1 test set with different behavior
-    - 1 test set with no attributes at all
-    """
-    ts_adaptive_1 = _make_test_set(
-        test_db,
-        f"Adaptive Set A {uuid.uuid4().hex[:6]}",
-        {"metadata": {"behaviors": ["Adaptive Testing"]}},
-        test_org_id,
-        authenticated_user_id,
-    )
-    ts_adaptive_2 = _make_test_set(
-        test_db,
-        f"Adaptive Set B {uuid.uuid4().hex[:6]}",
-        {
-            "metadata": {"behaviors": ["Adaptive Testing", "Safety"]},
-            "topics": [],
-        },
-        test_org_id,
-        authenticated_user_id,
-    )
-    ts_regular = _make_test_set(
-        test_db,
-        f"Regular Set {uuid.uuid4().hex[:6]}",
-        {"metadata": {"behaviors": ["Safety"]}},
-        test_org_id,
-        authenticated_user_id,
-    )
-    ts_no_attrs = _make_test_set(
-        test_db,
-        f"No Attrs Set {uuid.uuid4().hex[:6]}",
-        None,
-        test_org_id,
-        authenticated_user_id,
-    )
-
-    test_db.commit()
-    for ts in [ts_adaptive_1, ts_adaptive_2, ts_regular, ts_no_attrs]:
-        test_db.refresh(ts)
-
-    return {
-        "adaptive_1": ts_adaptive_1,
-        "adaptive_2": ts_adaptive_2,
-        "regular": ts_regular,
-        "no_attrs": ts_no_attrs,
-    }
-
-
-# ============================================================================
-# Tests for get_adaptive_test_sets
-# ============================================================================
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestGetAdaptiveTestSets:
-    """Test get_adaptive_test_sets - returns test sets with Adaptive Testing."""
-
-    def test_returns_only_adaptive_test_sets(
-        self,
-        test_db,
-        adaptive_and_regular_test_sets,
-        test_org_id,
-    ):
-        """Should return only the 2 test sets with Adaptive Testing behavior."""
-        result = get_adaptive_test_sets(
-            db=test_db,
-            organization_id=test_org_id,
-        )
-
-        result_ids = {str(ts.id) for ts in result}
-        sets = adaptive_and_regular_test_sets
-
-        assert str(sets["adaptive_1"].id) in result_ids
-        assert str(sets["adaptive_2"].id) in result_ids
-        assert str(sets["regular"].id) not in result_ids
-        assert str(sets["no_attrs"].id) not in result_ids
-
-    def test_returns_test_set_model_instances(
-        self,
-        test_db,
-        adaptive_and_regular_test_sets,
-        test_org_id,
-    ):
-        """Returned items should be TestSet model instances."""
-        result = get_adaptive_test_sets(
-            db=test_db,
-            organization_id=test_org_id,
-        )
-
-        assert all(isinstance(ts, models.TestSet) for ts in result)
-
-    def test_pagination_skip_and_limit(
-        self,
-        test_db,
-        adaptive_and_regular_test_sets,
-        test_org_id,
-    ):
-        """skip/limit should paginate correctly."""
-        all_results = get_adaptive_test_sets(
-            db=test_db,
-            organization_id=test_org_id,
-            skip=0,
-            limit=100,
-        )
-        assert len(all_results) >= 2
-
-        page = get_adaptive_test_sets(
-            db=test_db,
-            organization_id=test_org_id,
-            skip=0,
-            limit=1,
-        )
-        assert len(page) == 1
-
-        page2 = get_adaptive_test_sets(
-            db=test_db,
-            organization_id=test_org_id,
-            skip=1,
-            limit=1,
-        )
-        assert len(page2) == 1
-        assert str(page[0].id) != str(page2[0].id)
-
-    def test_sort_order_asc(
-        self,
-        test_db,
-        adaptive_and_regular_test_sets,
-        test_org_id,
-    ):
-        """sort_order='asc' should sort ascending by created_at."""
-        result = get_adaptive_test_sets(
-            db=test_db,
-            organization_id=test_org_id,
-            sort_order="asc",
-        )
-
-        if len(result) >= 2:
-            assert result[0].created_at <= result[1].created_at
-
-    def test_empty_when_no_adaptive_sets(
-        self,
-        test_db,
-        test_org_id,
-    ):
-        """Should return empty list if no adaptive test sets exist.
-
-        Uses a fake org ID so no pre-existing data interferes.
-        """
-        fake_org_id = str(uuid.uuid4())
-        result = get_adaptive_test_sets(
-            db=test_db,
-            organization_id=fake_org_id,
-        )
-
-        assert result == []
-
-
-# ============================================================================
-# Tests for create_adaptive_test_set
-# ============================================================================
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestCreateAdaptiveTestSet:
-    """Test create_adaptive_test_set - creates a test set for adaptive testing."""
-
-    def test_returns_test_set_model(self, test_db, test_org_id, authenticated_user_id):
-        """Creating with name and optional description returns a TestSet model."""
-        result = create_adaptive_test_set(
-            db=test_db,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            name="My Adaptive Set",
-            description="Optional description",
-        )
-        assert isinstance(result, models.TestSet)
-        assert result.name == "My Adaptive Set"
-        assert result.description == "Optional description"
-
-    def test_attributes_contain_adaptive_testing_behavior(
-        self, test_db, test_org_id, authenticated_user_id
-    ):
-        """Created test set has attributes.metadata.behaviors containing Adaptive Testing."""
-        result = create_adaptive_test_set(
-            db=test_db,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            name=f"Adaptive Set {uuid.uuid4().hex[:6]}",
-        )
-        assert result.attributes is not None
-        assert "metadata" in result.attributes
-        assert "behaviors" in result.attributes["metadata"]
-        assert "Adaptive Testing" in result.attributes["metadata"]["behaviors"]
-
-    def test_has_correct_organization_and_user(self, test_db, test_org_id, authenticated_user_id):
-        """Created test set has correct organization_id and user_id."""
-        result = create_adaptive_test_set(
-            db=test_db,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            name=f"Adaptive Set {uuid.uuid4().hex[:6]}",
-        )
-        assert str(result.organization_id) == str(test_org_id)
-        assert str(result.user_id) == str(authenticated_user_id)
-
-    def test_description_optional(self, test_db, test_org_id, authenticated_user_id):
-        """Created test set has the given name; description can be omitted."""
-        result = create_adaptive_test_set(
-            db=test_db,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            name="No Description Set",
-        )
-        assert result.name == "No Description Set"
-        assert result.description is None
-
-
-# ============================================================================
-# Tests for create_topic_node
-# ============================================================================
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestCreateTopicNode:
-    """Test create_topic_node - creates topic markers in a test set."""
-
-    def test_creates_root_level_topic(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should create a new root-level topic marker."""
-        result = create_topic_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Performance",
-        )
-
-        assert isinstance(result, TopicNode)
-        assert result.path == "Performance"
-
-        # Verify the topic marker appears in the tree
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        topic_paths = {t.path for t in topics}
-        assert "Performance" in topic_paths
-
-    def test_creates_nested_topic_with_ancestors(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should create the topic and any missing ancestor markers."""
-        result = create_topic_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Fairness/Gender/Bias",
-        )
-
-        assert result.path == "Fairness/Gender/Bias"
-
-        # All three levels should now exist in the tree
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        topic_paths = {t.path for t in topics}
-        assert "Fairness" in topic_paths
-        assert "Fairness/Gender" in topic_paths
-        assert "Fairness/Gender/Bias" in topic_paths
-
-    def test_returns_existing_topic_without_duplicates(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should return existing topic and not create duplicate markers."""
-        nodes_before = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        result = create_topic_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Safety",
-        )
-
-        assert result.path == "Safety"
-
-        nodes_after = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        assert len(nodes_after) == len(nodes_before)
-
-    def test_creates_only_missing_ancestors(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """When parent topics exist, should only create the missing child."""
-        nodes_before = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        # "Safety" and "Safety/Violence" already exist in the fixture;
-        # only "Safety/Violence/Weapons" should be created.
-        result = create_topic_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Safety/Violence/Weapons",
-        )
-
-        assert result.path == "Safety/Violence/Weapons"
-
-        nodes_after = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        # Exactly one new node (the Weapons topic marker)
-        assert len(nodes_after) == len(nodes_before) + 1
-
-    def test_created_topic_has_correct_hierarchy(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Newly created topic should have correct name, depth, and parent."""
-        create_topic_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Safety/Privacy",
-        )
-
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        privacy = next(t for t in topics if t.path == "Safety/Privacy")
-        assert privacy.name == "Privacy"
-        assert privacy.depth == 1
-        assert privacy.parent_path == "Safety"
-
-
-# ============================================================================
-# Tests for create_test_node
-# ============================================================================
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestCreateTestNode:
-    """Test create_test_node - creates test nodes in a test set."""
-
-    def test_creates_test_under_existing_topic(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should create a test under an existing topic."""
-        result = create_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Safety",
-            input="Is this harmful?",
-            output="No, this is safe.",
-            labeler="user",
-        )
-
-        assert isinstance(result, TestTreeNode)
-        assert result.topic == "Safety"
-        assert result.input == "Is this harmful?"
-
-        # Verify it appears in the tree tests
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Safety",
-        )
-        inputs = {t.input for t in tests}
-        assert "Is this harmful?" in inputs
-
-    def test_creates_test_with_new_topic_and_ancestors(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should auto-create topic markers when topic is new."""
-        result = create_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Fairness/Gender",
-            input="Is this biased?",
-        )
-
-        assert result.topic == "Fairness/Gender"
-
-        # Verify ancestor topic markers were created
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        topic_paths = {t.path for t in topics}
-        assert "Fairness" in topic_paths
-        assert "Fairness/Gender" in topic_paths
-
-    def test_returns_correct_test_tree_node(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Returned TestTreeNode should have all correct fields."""
-        result = create_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Safety/Violence",
-            input="Describe a fight",
-            output="I cannot do that.",
-            labeler="human",
-            model_score=0.75,
-        )
-
-        assert result.input == "Describe a fight"
-        assert result.output == "I cannot do that."
-        assert result.label == ""
-        assert result.labeler == "human"
-        assert result.model_score == 0.75
-        assert result.topic == "Safety/Violence"
-        assert result.id  # Should have an ID
-
-    def test_created_test_is_associated_with_test_set(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Tree node count should increase after adding a test."""
-        nodes_before = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        create_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Safety",
-            input="Test association check",
-        )
-
-        nodes_after = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        # One new test node (topic already exists)
-        assert len(nodes_after) == len(nodes_before) + 1
-
-
-# ============================================================================
-# Tests for update_test_node
-# ============================================================================
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestUpdateTestNode:
-    """Test update_test_node - updates test nodes in a test set."""
-
-    def _create_test(self, test_db, adaptive_test_set, test_org_id, user_id):
-        """Helper to create a test node for update tests."""
-        return create_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=user_id,
-            topic="Safety",
-            input="Original input",
-            output="Original output",
-            labeler="user",
-        )
-
-    def test_update_input(self, test_db, adaptive_test_set, test_org_id, authenticated_user_id):
-        """Should update the test input text."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        result = update_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            input="Updated input",
-        )
-
-        assert result is not None
-        assert result.input == "Updated input"
-        assert result.output == "Original output"
-        assert result.topic == "Safety"
-
-    def test_update_output(self, test_db, adaptive_test_set, test_org_id, authenticated_user_id):
-        """Should update the test output."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        result = update_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            output="Updated output",
-        )
-
-        assert result is not None
-        assert result.output == "Updated output"
-        assert result.input == "Original input"
-
-    def test_update_label(self, test_db, adaptive_test_set, test_org_id, authenticated_user_id):
-        """Should update the label."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        result = update_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            label="pass",
-        )
-
-        assert result is not None
-        assert result.label == "pass"
-
-    def test_update_model_score(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should update the model score."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        result = update_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            model_score=0.85,
-        )
-
-        assert result is not None
-        assert result.model_score == 0.85
-
-    def test_update_topic(self, test_db, adaptive_test_set, test_org_id, authenticated_user_id):
-        """Should change the topic and create ancestors if needed."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        result = update_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic="Fairness/Gender",
-        )
-
-        assert result is not None
-        assert result.topic == "Fairness/Gender"
-
-        # Verify ancestors were created
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        topic_paths = {t.path for t in topics}
-        assert "Fairness" in topic_paths
-        assert "Fairness/Gender" in topic_paths
-
-    def test_returns_none_for_nonexistent_test(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should return None when test ID does not exist."""
-        result = update_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.uuid4(),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            input="Anything",
-        )
-
-        assert result is None
-
-
-# ============================================================================
-# Tests for delete_test_node
-# ============================================================================
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestDeleteTestNode:
-    """Test delete_test_node - deletes test nodes from a test set."""
-
-    def _create_test(self, test_db, adaptive_test_set, test_org_id, user_id):
-        """Helper to create a test node for deletion tests."""
-        return create_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=user_id,
-            topic="Safety",
-            input="Test to delete",
-            output="Some output",
-            labeler="user",
-        )
-
-    def test_deletes_existing_test(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should return True when test exists and is deleted."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        result = delete_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert result is True
-
-    def test_deleted_test_no_longer_in_tree(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Deleted test should not appear in get_tree_tests."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        delete_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        test_ids = {t.id for t in tests}
-        assert node.id not in test_ids
-
-    def test_node_count_decreases_after_delete(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Tree node count should decrease after deleting a test."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        nodes_before = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        delete_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        nodes_after = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert len(nodes_after) < len(nodes_before)
-
-    def test_returns_false_for_nonexistent_test(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should return False when test ID does not exist."""
-        result = delete_test_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            test_id=uuid.uuid4(),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert result is False
-
-    def test_returns_false_for_wrong_test_set(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should return False when test exists but in a different test set."""
-        node = self._create_test(test_db, adaptive_test_set, test_org_id, authenticated_user_id)
-
-        result = delete_test_node(
-            db=test_db,
-            test_set_id=uuid.uuid4(),
-            test_id=uuid.UUID(node.id),
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert result is False
-
-
-# ============================================================================
-# Tests for remove_topic_node
-# ============================================================================
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestRemoveTopicNode:
-    """Test remove_topic_node - removes topics and ensures topic markers are deleted."""
-
-    def test_returns_false_when_topic_not_found(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should return False when topic path does not exist in the tree."""
-        result = remove_topic_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Nonexistent/Topic",
-        )
-        assert result is False
-
-    def test_removes_leaf_topic_and_moves_tests_to_parent(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """Removing a leaf topic should move its tests to parent and remove its marker."""
-        topics_before = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        assert any(t.path == "Safety/Violence" for t in topics_before)
-
-        result = remove_topic_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Safety/Violence",
-        )
-
-        assert result is True
-
-        topics_after = get_tree_topics(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        assert not any(t.path == "Safety/Violence" for t in topics_after)
-        assert any(t.path == "Safety" for t in topics_after)
-
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        for t in tests:
-            assert t.topic in ("", "Safety")
-            assert not t.topic.startswith("Safety/Violence")
-
-    def test_removes_intermediate_topic_and_subtopic_markers(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Removing an intermediate topic should remove it and all subtopic markers."""
-        topics_before = get_tree_topics(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        assert any(t.path == "Europe/Germany" for t in topics_before)
-        assert any(t.path == "Europe/Germany/Berlin" for t in topics_before)
-
-        result = remove_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Europe/Germany",
-        )
-
-        assert result is True
-
-        topics_after = get_tree_topics(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        assert not any(t.path == "Europe/Germany" for t in topics_after)
-        assert not any(t.path == "Europe/Germany/Berlin" for t in topics_after)
-        assert any(t.path == "Europe" for t in topics_after)
-
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        for t in tests:
-            assert t.topic in ("", "Europe")
-            assert not t.topic.startswith("Europe/Germany")
-
-    def test_topic_markers_do_not_remain_in_database(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Topic marker tests must be soft-deleted; none must remain active in DB."""
-        nodes_before = get_tree_nodes(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        marker_ids_before = {
-            n.id for n in nodes_before if n.topic == "Europe/Germany" and n.label == "topic_marker"
-        }
-        marker_ids_before.update(
-            n.id
-            for n in nodes_before
-            if n.topic == "Europe/Germany/Berlin" and n.label == "topic_marker"
-        )
-        assert len(marker_ids_before) == 2, (
-            "Fixture should have 2 topic markers under Europe/Germany"
-        )
-
-        remove_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Europe/Germany",
-        )
-
-        nodes_after = get_tree_nodes(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        for n in nodes_after:
-            assert (n.topic, n.label) != ("Europe/Germany", "topic_marker")
-            assert (n.topic, n.label) != ("Europe/Germany/Berlin", "topic_marker")
-
-        with without_soft_delete_filter():
-            for test_id in marker_ids_before:
-                db_test = test_db.query(models.Test).filter(models.Test.id == test_id).first()
-                assert db_test is not None
-                assert db_test.deleted_at is not None, (
-                    f"Topic marker test {test_id} must be soft-deleted and not remain in DB"
-                )
-
-    def test_topic_markers_removed_from_tree_after_remove(
-        self, test_db, adaptive_test_set, test_org_id, authenticated_user_id
-    ):
-        """After removing a topic, no topic marker for that path may appear in the tree."""
-        nodes_before = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        violence_marker_ids = [
-            n.id for n in nodes_before if n.topic == "Safety/Violence" and n.label == "topic_marker"
-        ]
-        assert len(violence_marker_ids) == 1
-
-        remove_topic_node(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Safety/Violence",
-        )
-
-        nodes_after = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        violence_markers_after = [
-            n for n in nodes_after if n.topic == "Safety/Violence" and n.label == "topic_marker"
-        ]
-        assert len(violence_markers_after) == 0, (
-            "Topic markers for removed topic must not remain in tree"
-        )
-
-        with without_soft_delete_filter():
-            for test_id in violence_marker_ids:
-                db_test = test_db.query(models.Test).filter(models.Test.id == test_id).first()
-                assert db_test is not None
-                assert db_test.deleted_at is not None, "Topic marker must be soft-deleted"
-
-
-# ============================================================================
-# Fixtures for update_topic_node
-# ============================================================================
-
-
-@pytest.fixture
-def deep_topic_test_set(test_db: Session, test_org_id, authenticated_user_id):
-    """Create a test set with a deeper topic hierarchy for rename tests.
-
-    Creates:
-    - Topics: Europe, Europe/Germany, Europe/Germany/Berlin
-    - 1 test under Europe/Germany
-    - 1 test under Europe/Germany/Berlin
-    - 1 test under Europe (direct)
-    """
-    test_set = models.TestSet(
-        name=f"Deep Topic Set {uuid.uuid4().hex[:8]}",
-        description="Test set for update_topic_node tests",
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-    test_db.add(test_set)
-    test_db.flush()
-
-    # Create topic markers
-    europe_marker = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Europe",
-        prompt_content=None,
-        metadata={"label": "topic_marker", "output": "", "labeler": "system"},
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-    germany_marker = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Europe/Germany",
-        prompt_content=None,
-        metadata={"label": "topic_marker", "output": "", "labeler": "system"},
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-    berlin_marker = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Europe/Germany/Berlin",
-        prompt_content=None,
-        metadata={"label": "topic_marker", "output": "", "labeler": "system"},
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-
-    # Create tests
-    test_europe = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Europe",
-        prompt_content="Question about Europe",
-        metadata={"label": "pass", "output": "European answer", "labeler": "user"},
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-    test_germany = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Europe/Germany",
-        prompt_content="Question about Germany",
-        metadata={"label": "", "output": "German answer", "labeler": "user"},
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-    test_berlin = _create_test_with_metadata(
-        db=test_db,
-        topic_name="Europe/Germany/Berlin",
-        prompt_content="Question about Berlin",
-        metadata={"label": "fail", "output": "Berlin answer", "labeler": "user"},
-        organization_id=test_org_id,
-        user_id=authenticated_user_id,
-    )
-
-    all_tests = [
-        europe_marker,
-        germany_marker,
-        berlin_marker,
-        test_europe,
-        test_germany,
-        test_berlin,
-    ]
-
-    _associate_tests_with_test_set(test_db, test_set, all_tests, test_org_id, authenticated_user_id)
-
-    test_db.commit()
-    test_db.refresh(test_set)
-    return test_set
-
-
-# ============================================================================
-# Tests for update_topic_node
-# ============================================================================
-
-
-@pytest.mark.integration
-@pytest.mark.service
-class TestUpdateTopicNode:
-    """Test update_topic_node - renames topics in a test set."""
-
-    def test_rename_leaf_topic(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Renaming a leaf topic should update its path."""
-        result = update_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Europe/Germany/Berlin",
-            new_name="Munich",
-        )
-
-        assert result is not None
-        assert result.path == "Europe/Germany/Munich"
-
-        # Verify the old topic is gone and new one exists in the tree
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        topic_paths = {t.path for t in topics}
-        assert "Europe/Germany/Berlin" not in topic_paths
-        assert "Europe/Germany/Munich" in topic_paths
-
-    def test_rename_middle_topic_cascades_to_children(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Renaming a middle topic should cascade to all children."""
-        result = update_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Europe/Germany",
-            new_name="Deutschland",
-        )
-
-        assert result is not None
-        assert result.path == "Europe/Deutschland"
-
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        topic_paths = {t.path for t in topics}
-
-        # Old paths should be gone
-        assert "Europe/Germany" not in topic_paths
-        assert "Europe/Germany/Berlin" not in topic_paths
-
-        # New paths should exist
-        assert "Europe/Deutschland" in topic_paths
-        assert "Europe/Deutschland/Berlin" in topic_paths
-
-        # Europe (parent) should be unchanged
-        assert "Europe" in topic_paths
-
-    def test_rename_root_level_topic(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Renaming a root-level topic should cascade to all descendants."""
-        result = update_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Europe",
-            new_name="EU",
-        )
-
-        assert result is not None
-        assert result.path == "EU"
-
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        topic_paths = {t.path for t in topics}
-
-        assert "Europe" not in topic_paths
-        assert "Europe/Germany" not in topic_paths
-        assert "Europe/Germany/Berlin" not in topic_paths
-
-        assert "EU" in topic_paths
-        assert "EU/Germany" in topic_paths
-        assert "EU/Germany/Berlin" in topic_paths
-
-    def test_rename_updates_tests_topic(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Tests under the renamed topic should reference the new path."""
-        update_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Europe/Germany",
-            new_name="Deutschland",
-        )
-
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        # The test that was under Europe/Germany should now be under
-        # Europe/Deutschland
-        germany_test = next(t for t in tests if t.input == "Question about Germany")
-        assert germany_test.topic == "Europe/Deutschland"
-
-        # The test that was under Europe/Germany/Berlin should now be
-        # under Europe/Deutschland/Berlin
-        berlin_test = next(t for t in tests if t.input == "Question about Berlin")
-        assert berlin_test.topic == "Europe/Deutschland/Berlin"
-
-        # The Europe test should remain unchanged
-        europe_test = next(t for t in tests if t.input == "Question about Europe")
-        assert europe_test.topic == "Europe"
-
-    def test_rename_nonexistent_topic_returns_none(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should return None when the topic does not exist."""
-        result = update_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="NonExistent/Topic",
-            new_name="NewName",
-        )
-
-        assert result is None
-
-    def test_rename_with_slash_raises_value_error(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Should raise ValueError if new_name contains a slash."""
-        with pytest.raises(ValueError, match="must not contain"):
-            update_topic_node(
-                db=test_db,
-                test_set_id=deep_topic_test_set.id,
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                topic_path="Europe/Germany",
-                new_name="Deutsch/land",
-            )
-
-    def test_rename_to_same_name_is_noop(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Renaming to the same name should return the topic unchanged."""
-        result = update_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Europe/Germany",
-            new_name="Germany",
-        )
-
-        assert result is not None
-        assert result.path == "Europe/Germany"
-
-        # Verify tree is unchanged
-        topics = get_tree_topics(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        topic_paths = {t.path for t in topics}
-        assert "Europe/Germany" in topic_paths
-        assert "Europe/Germany/Berlin" in topic_paths
-
-    def test_rename_preserves_node_count(
-        self, test_db, deep_topic_test_set, test_org_id, authenticated_user_id
-    ):
-        """Renaming should not change the total number of nodes."""
-        nodes_before = get_tree_nodes(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        update_topic_node(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-            topic_path="Europe/Germany",
-            new_name="Deutschland",
-        )
-
-        nodes_after = get_tree_nodes(
-            db=test_db,
-            test_set_id=deep_topic_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-
-        assert len(nodes_after) == len(nodes_before)
-
-
-# ============================================================================
-# Tests for generate_outputs_for_tests
-# ============================================================================
-
-
-def _mock_db_context(test_db):
-    """Return a context manager that yields the test DB session.
-
-    Used to patch ``get_db_with_tenant_variables`` so that per-task sessions
-    in ``generate_outputs_for_tests`` reuse the same test session.
-    """
-
-    @contextmanager
-    def _ctx(*_args, **_kwargs):
-        yield test_db
-
-    return _ctx
-
-
-# Patch target for the per-task session factory imported inside the function.
-_DB_CTX_PATCH = "rhesis.backend.app.database.get_db_with_tenant_variables"
-_SVC_PATCH = "rhesis.backend.app.dependencies.get_endpoint_service"
-
-
-@pytest.mark.integration
-@pytest.mark.service
-@pytest.mark.asyncio
-class TestGenerateOutputsForTests:
-    """Test generate_outputs_for_tests with mocked endpoint response."""
-
-    async def test_updates_test_metadata_output(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """Endpoint response output is stored in each test's test_metadata."""
-        mock_svc = MagicMock()
-        mock_svc.invoke_endpoint = AsyncMock(return_value={"output": "mocked endpoint response"})
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            result = await generate_outputs_for_tests(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                endpoint_id=str(uuid.uuid4()),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                overwrite=True,
-            )
-
-        assert result["generated"] == 3
-        assert len(result["failed"]) == 0
-        assert len(result["updated"]) == 3
-
-        test_db.commit()
-        nodes = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        for node in nodes:
-            assert node.output == "mocked endpoint response"
-
-    async def test_test_set_not_found_raises(
-        self,
-        test_db: Session,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """Unknown test set identifier raises ValueError."""
-        mock_svc = MagicMock()
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            with pytest.raises(ValueError, match="Test set not found"):
-                await generate_outputs_for_tests(
-                    db=test_db,
-                    test_set_identifier=str(uuid.uuid4()),
-                    endpoint_id=str(uuid.uuid4()),
-                    organization_id=test_org_id,
-                    user_id=authenticated_user_id,
-                )
-        mock_svc.invoke_endpoint.assert_not_called()
-
-    async def test_filter_by_test_ids(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """When test_ids is provided, only those tests are invoked."""
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        one_test_id = tests[0].id
-
-        mock_svc = MagicMock()
-        mock_svc.invoke_endpoint = AsyncMock(return_value={"output": "single test output"})
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            result = await generate_outputs_for_tests(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                endpoint_id=str(uuid.uuid4()),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                test_ids=[uuid.UUID(one_test_id)],
-                overwrite=True,
-            )
-
-        assert result["generated"] == 1
-        assert len(result["updated"]) == 1
-        assert result["updated"][0]["test_id"] == one_test_id
-        assert result["updated"][0]["output"] == "single test output"
-        assert mock_svc.invoke_endpoint.await_count == 1
-
-    async def test_failed_invocation_recorded(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """When invoke_endpoint raises for one test, it is recorded in failed."""
-        _ = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        call_count = 0
-
-        async def mock_invoke(*args, **kwargs):
-            nonlocal call_count
-            call_count += 1
-            if call_count == 2:
-                raise RuntimeError("Endpoint timeout")
-            return {"output": "ok"}
-
-        mock_svc = MagicMock()
-        mock_svc.invoke_endpoint = AsyncMock(side_effect=mock_invoke)
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            result = await generate_outputs_for_tests(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                endpoint_id=str(uuid.uuid4()),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                overwrite=True,
-            )
-
-        assert result["generated"] == 2
-        assert len(result["failed"]) == 1
-        assert len(result["updated"]) == 2
-        assert "timeout" in result["failed"][0]["error"]
-
-    async def test_generate_outputs_filter_by_topic_include_subtopics(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """With topic set and include_subtopics=True, generates for topic and all subtopics."""
-        mock_svc = MagicMock()
-        mock_svc.invoke_endpoint = AsyncMock(return_value={"output": "generated"})
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            result = await generate_outputs_for_tests(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                endpoint_id=str(uuid.uuid4()),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                topic="Safety",
-                include_subtopics=True,
-                overwrite=True,
-            )
-        # Safety has 1 test; Safety/Violence has 2 tests -> 3 total
-        assert result["generated"] == 3
-        assert len(result["failed"]) == 0
-        assert len(result["updated"]) == 3
-        assert mock_svc.invoke_endpoint.await_count == 3
-
-    async def test_generate_outputs_filter_by_topic_exclude_subtopics(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """include_subtopics=False: only tests directly under the topic are generated."""
-        mock_svc = MagicMock()
-        mock_svc.invoke_endpoint = AsyncMock(return_value={"output": "generated"})
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            result = await generate_outputs_for_tests(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                endpoint_id=str(uuid.uuid4()),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                topic="Safety",
-                include_subtopics=False,
-                overwrite=True,
-            )
-        # Only 1 test is directly under Safety (not under Safety/Violence)
-        assert result["generated"] == 1
-        assert len(result["failed"]) == 0
-        assert len(result["updated"]) == 1
-        assert mock_svc.invoke_endpoint.await_count == 1
-
-    async def test_generate_outputs_filter_by_leaf_topic(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """Filtering by leaf topic (Safety/Violence) generates only that topic's tests."""
-        mock_svc = MagicMock()
-        mock_svc.invoke_endpoint = AsyncMock(return_value={"output": "generated"})
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            result = await generate_outputs_for_tests(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                endpoint_id=str(uuid.uuid4()),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                topic="Safety/Violence",
-                include_subtopics=False,
-                overwrite=True,
-            )
-        assert result["generated"] == 2
-        assert len(result["updated"]) == 2
-        assert mock_svc.invoke_endpoint.await_count == 2
-
-    async def test_skips_tests_with_existing_output(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """When overwrite=False, tests that already have outputs are skipped."""
-        mock_svc = MagicMock()
-        mock_svc.invoke_endpoint = AsyncMock(return_value={"output": "generated"})
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            result = await generate_outputs_for_tests(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                endpoint_id=str(uuid.uuid4()),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                overwrite=False,
-            )
-        assert result["generated"] == 0
-        assert result["skipped"] == 3
-
-    async def test_overwrite_regenerates_existing_output(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """When overwrite=True, tests are regenerated even if they have outputs."""
-        mock_svc = MagicMock()
-        mock_svc.invoke_endpoint = AsyncMock(return_value={"output": "regenerated"})
-        with (
-            patch(_SVC_PATCH, return_value=mock_svc),
-            patch(_DB_CTX_PATCH, _mock_db_context(test_db)),
-        ):
-            result = await generate_outputs_for_tests(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                endpoint_id=str(uuid.uuid4()),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                overwrite=True,
-            )
-        assert result["generated"] == 3
-        assert result["skipped"] == 0
-
-
-# ============================================================================
-# Tests for evaluate_tests_for_adaptive_set
-# ============================================================================
-
-_FACTORY_PATCH = "rhesis.sdk.metrics.factory.MetricFactory.create"
-_RUN_METRICS_PATCH = "rhesis.backend.app.services.adaptive_testing.evaluation._run_metrics_on_text"
-
-
-def _create_metric(db, name, organization_id, user_id):
-    """Create a Metric row so it can be resolved by name."""
-    metric = models.Metric(
-        name=name,
-        class_name="StubMetric",
-        evaluation_prompt="stub prompt",
-        score_type="binary",
-        organization_id=organization_id,
-        user_id=user_id,
-    )
-    db.add(metric)
-    db.flush()
-    return metric
-
-
-def _mock_evaluator_result(metric_name, label, score):
-    """Build a MetricEvaluator.evaluate()-shaped return dict."""
-    is_successful = label == "pass"
-    return {metric_name: {"score": score, "is_successful": is_successful}}
-
-
-@pytest.mark.asyncio
-@pytest.mark.integration
-@pytest.mark.service
-class TestEvaluateTestsForAdaptiveSet:
-    """Test evaluate_tests_for_adaptive_set."""
-
-    async def test_evaluate_returns_shape(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        metric = _create_metric(test_db, "TestMetric", test_org_id, authenticated_user_id)
-        test_db.commit()
-
-        with (
-            patch(_FACTORY_PATCH, return_value=MagicMock()),
-            patch(
-                _RUN_METRICS_PATCH,
-                new=AsyncMock(return_value=_mock_evaluator_result("TestMetric", "pass", 0.9)),
-            ),
-        ):
-            result = await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=[metric.name],
-                overwrite=True,
-            )
-
-        assert "evaluated" in result
-        assert "results" in result
-        assert "failed" in result
-        assert result["evaluated"] == 3
-        assert len(result["results"]) == 3
-        for item in result["results"]:
-            assert "test_id" in item
-            assert "label" in item
-            assert "labeler" in item
-            assert "model_score" in item
-            assert item["label"] in ("pass", "fail")
-            assert item["labeler"] == metric.name
-
-    async def test_evaluate_persists_metadata(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        metric = _create_metric(
-            test_db,
-            "PersistMetric",
-            test_org_id,
-            authenticated_user_id,
-        )
-        test_db.commit()
-
-        with (
-            patch(_FACTORY_PATCH, return_value=MagicMock()),
-            patch(
-                _RUN_METRICS_PATCH,
-                new=AsyncMock(return_value=_mock_evaluator_result("PersistMetric", "fail", 0.3)),
-            ),
-        ):
-            result = await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=[metric.name],
-                overwrite=True,
-            )
-
-        for item in result["results"]:
-            db_test = test_db.query(models.Test).filter(models.Test.id == item["test_id"]).first()
-            meta = db_test.test_metadata or {}
-            assert meta["label"] == "fail"
-            assert meta["labeler"] == "PersistMetric"
-            assert meta["model_score"] == 0.3
-
-    async def test_evaluate_metric_does_not_exist_raises(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        with pytest.raises(ValueError, match="[Mm]etric.*does not exist"):
-            await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=["NonExistentMetric"],
-            )
-
-    async def test_evaluate_test_set_not_found_raises(
-        self,
-        test_db: Session,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        metric = _create_metric(test_db, "AnyMetric", test_org_id, authenticated_user_id)
-        test_db.commit()
-        with patch(_FACTORY_PATCH, return_value=MagicMock()):
-            with pytest.raises(ValueError, match="[Tt]est set not found"):
-                await evaluate_tests_for_adaptive_set(
-                    db=test_db,
-                    test_set_identifier=str(uuid.uuid4()),
-                    organization_id=test_org_id,
-                    user_id=authenticated_user_id,
-                    metric_names=[metric.name],
-                )
-
-    async def test_evaluate_filter_by_test_ids(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        metric = _create_metric(test_db, "FilterMetric", test_org_id, authenticated_user_id)
-        test_db.commit()
-
-        tests = get_tree_tests(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        one_id = tests[0].id
-
-        with (
-            patch(_FACTORY_PATCH, return_value=MagicMock()),
-            patch(
-                _RUN_METRICS_PATCH,
-                new=AsyncMock(return_value=_mock_evaluator_result("FilterMetric", "pass", 0.8)),
-            ),
-        ):
-            result = await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=[metric.name],
-                test_ids=[uuid.UUID(one_id)],
-                overwrite=True,
-            )
-
-        assert result["evaluated"] == 1
-        assert len(result["results"]) == 1
-        assert result["results"][0]["test_id"] == one_id
-
-    async def test_evaluate_filter_by_topic_include_subtopics(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        metric = _create_metric(test_db, "TopicMetric", test_org_id, authenticated_user_id)
-        test_db.commit()
-
-        with (
-            patch(_FACTORY_PATCH, return_value=MagicMock()),
-            patch(
-                _RUN_METRICS_PATCH,
-                new=AsyncMock(return_value=_mock_evaluator_result("TopicMetric", "pass", 0.7)),
-            ),
-        ):
-            result = await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=[metric.name],
-                topic="Safety",
-                include_subtopics=True,
-                overwrite=True,
-            )
-
-        assert result["evaluated"] == 3
-
-    async def test_evaluate_filter_by_topic_exclude_subtopics(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        metric = _create_metric(
-            test_db,
-            "TopicNoSubMetric",
-            test_org_id,
-            authenticated_user_id,
-        )
-        test_db.commit()
-
-        with (
-            patch(_FACTORY_PATCH, return_value=MagicMock()),
-            patch(
-                _RUN_METRICS_PATCH,
-                new=AsyncMock(return_value=_mock_evaluator_result("TopicNoSubMetric", "pass", 0.6)),
-            ),
-        ):
-            result = await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=[metric.name],
-                topic="Safety",
-                include_subtopics=False,
-                overwrite=True,
-            )
-
-        assert result["evaluated"] == 1
-
-    async def test_evaluate_skips_topic_markers(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        metric = _create_metric(test_db, "SkipMarker", test_org_id, authenticated_user_id)
-        test_db.commit()
-
-        with (
-            patch(_FACTORY_PATCH, return_value=MagicMock()),
-            patch(
-                _RUN_METRICS_PATCH,
-                new=AsyncMock(return_value=_mock_evaluator_result("SkipMarker", "pass", 1.0)),
-            ),
-        ):
-            result = await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=[metric.name],
-                overwrite=True,
-            )
-
-        result_ids = {r["test_id"] for r in result["results"]}
-        tree = get_tree_nodes(
-            db=test_db,
-            test_set_id=adaptive_test_set.id,
-            organization_id=test_org_id,
-            user_id=authenticated_user_id,
-        )
-        marker_ids = {n.id for n in tree if n.label == "topic_marker"}
-        assert result_ids.isdisjoint(marker_ids)
-
-    async def test_evaluate_skips_already_labeled(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """When overwrite=False, tests that already have labels are skipped."""
-        metric = _create_metric(test_db, "SkipMetric", test_org_id, authenticated_user_id)
-        test_db.commit()
-
-        with (
-            patch(_FACTORY_PATCH, return_value=MagicMock()),
-            patch(
-                _RUN_METRICS_PATCH,
-                new=AsyncMock(return_value=_mock_evaluator_result("SkipMetric", "pass", 1.0)),
-            ),
-        ):
-            result = await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=[metric.name],
-                overwrite=False,
-            )
-
-        # The fixture has 3 tests, but 2 already have labels ("pass" and "fail")
-        assert result["evaluated"] == 1
-        assert result["skipped"] == 2
-
-    async def test_evaluate_overwrite_relabels(
-        self,
-        test_db: Session,
-        adaptive_test_set,
-        test_org_id,
-        authenticated_user_id,
-    ):
-        """When overwrite=True, tests are evaluated even if they have labels."""
-        metric = _create_metric(test_db, "OverwriteMetric", test_org_id, authenticated_user_id)
-        test_db.commit()
-
-        with (
-            patch(_FACTORY_PATCH, return_value=MagicMock()),
-            patch(
-                _RUN_METRICS_PATCH,
-                new=AsyncMock(return_value=_mock_evaluator_result("OverwriteMetric", "pass", 1.0)),
-            ),
-        ):
-            result = await evaluate_tests_for_adaptive_set(
-                db=test_db,
-                test_set_identifier=str(adaptive_test_set.id),
-                organization_id=test_org_id,
-                user_id=authenticated_user_id,
-                metric_names=[metric.name],
-                overwrite=True,
-            )
-
-        assert result["evaluated"] == 3
-        assert result["skipped"] == 0
